Swapped awk/grep approach for seqkit/grep approach. Added verbosity.
Added an explanation on way awk approach may not be the best. It could work if the awk program could reverse the line breakage afterwards,
returning the FASTA file to its normal structure. When alignment was performed using both approaches, different base scores were obtained.
This clarifies that, although decontaminants.fastq file is indeed filtered, the deformation of its structure unallows STAR to recognize the
file, thus not performing the alignment accordingly.
Rest is just verbosity for clarification of the performed tasks